+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Added getPercentage util to numbers subset
 - Added CodePushStatus container tests
+- Added checked in icon to Place cards
 
 ### Changed